React front deploy with spring-boot + use create-react-app
