--- sources/platform/server/src/main/frontend/.babelrc
@@ -1,9 +1,0 @@
-{
-    "presets": [
-        "@babel/preset-env",
-        "@babel/preset-react"
-    ],
-    "plugins": [
-        "@babel/plugin-proposal-class-properties"
-    ]
-}
