+++ sources/platform/server/src/main/frontend/package.json
@@ -1,53 +1,31 @@
 {
-  "name": "frontend",
-  "version": "1.0.0",
-  "description": "Starter app ",
-  "main": "index.js",
-  "scripts": {
-    "build": "webpack",
-    "start": "webpack-dev-server --hot --inline",
-    "eslint": "eslint app"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/dlizarra/spring-boot-react-webpack-starter.git"
-  },
-  "license": "ISC",
-  "devDependencies": {
-    "@babel/core": "^7.1.0",
-    "@babel/preset-env": "^7.1.0",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.1.0",
-    "babel-loader": "^8.0.4",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-preset-react": "^6.5.0",
-    "css-loader": "^1.0.0",
-    "eslint": "^5.7.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-react": "^7.11.1",
-    "eslint-plugin-import": "^2.14.0",
-    "style-loader": "^0.23.1",
-    "webpack": "^4.21.0",
-    "webpack-dev-server": "^3.1.9",
-    "webpack-hot-middleware": "^2.24.3",
-    "webpack-merge": "^4.1.4",
-    "webpack-cli": "^3.1.2",
-    "winston": "^3.1.0"
-  },
+  "name": "pcu-front",
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "babel-preset-stage-1": "^6.5.0",
-    "history": "^4.7.2",
-    "lodash": "^4.17.11",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
-    "react-redux": "^5.0.7",
-    "react-router": "^4.3.1",
-    "react-router-bootstrap": "^0.24.4",
-    "react-router-redux": "^4.0.4",
-    "redux": "^4.0.1",
-    "redux-simple-router": "^2.0.4",
-    "redux-thunk": "^2.0.1",
+    "react-scripts": "2.0.5",
+    "axios": "^0.9.1",
     "moment": "^2.19.1"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ],
+  "devDependencies": {
+    "renamer": "^1.0.0",
+    "replace-in-file": "^3.4.2"
   }
 }
