+++ sources/platform/server/src/main/frontend/src/App.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react';
+import './style/main.css'
+
+import Search from './scenes/Search/Search.js';
+
+class App extends Component {
+  
+  render() {
+    return (
+        <Search />
+    );
+  }
+}
+
+export default App;
