+++ sources/platform/server/src/main/frontend/src/index.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import './index.css';
+import App from './App';
+import * as serviceWorker from './serviceWorker';
+
+//ReactDOM.render(<App />, document.getElementById('root'));
+
+ReactDOM.render(
+    <App/>,
+    document.getElementById('app'));
+
+serviceWorker.unregister();
