+++ sources/platform/server/src/main/frontend/src/style/components/_buttons.scss
