+++ sources/platform/server/src/main/frontend/src/style/components/_mobile.scss
