+++ sources/platform/server/src/main/frontend/src/style/main.scss
@@ -1,0 +1,10 @@
+@import "components/mixins";
+
+@import "components/general";
+@import "components/login";
+
+@import "components/mobile";
+@import "components/buttons";
+
+# social network icons :
+@import "font-awesome/font-awesome";
