--- sources/platform/server/src/main/resources/static/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1">    
-
-    <!-- Bootstrap CSS Sheet -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/css/bootstrap.min.css">
-    <!-- Google Fonts CDN -->
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Lato">
-
-</head>
-<body>
-    <div id="app" class="app"></div>
-    <script src="/bundle.js"></script>
-</body>
-</html>
